Rename module path
