+++ aworld/__main__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # Copyright (c) 2025 inclusionAI.
 
-from .command.cli import main
+from .cmd.cli import main
 
 if __name__ == "__main__":
     main()
